requirements.txt
pandas>=2.0.0
numpy>=1.24.0
seaborn>=0.12.0
matplotlib>=3.7.0
plotly>=5.14.0
jupyter>=1.0.0
jupyterlab>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
scipy>=1.10.0
pyarrow>=12.0.0
kaleido>=0.2.1

